+++ data-raw/merge-mspcts.r
@@ -1,6 +1,7 @@
 rm(list = ls(pattern = "*"))
 
 library(photobiology)
+library(photobiologyWavebands)
 
 # merge
 
@@ -28,36 +29,41 @@ lamps.mspct <- c(licor.mspct, macam.mspct, bentham.mspct, oo_maya.mspct,
 
 lamps.mspct <- lamps.mspct[order(names(lamps.mspct))]
 
-uv_lamps <- 
-        sort(c("germicidal", "philips.tl01", "philips.tl12", 
-               "philips.tl12.mc", "philips.BLB.tld108", "LCDlighting.UVA.BL",
-               "qpanel.uva340", "qpanel.uvb313", "convoy.2p.uv.flashlight",
-               "cree.uv.flashlight", "jaxman.uc1flood.uv.flashlight"))
-amber_lamps <- sodium_lamps <- c("osram.super.vialox")
-red_lamps <- "philips.tld36w.15"
-blue_lamps <- c("philips.tld36w.18", "philips.tld.36w.18")
-white_lamps <- 
-        sort(c("philips.pls11w.827", "philips.tld36w.83", 
-               "philips.tld36w.865",
-               "philips.tld36w.89",  "philips.tld36w.92",  "philips.tll36w.950",
-               "philips.led.t8.10w.840",
-               "sylvania.215w.vho", 
-               "airam.cf.15w.2700k", "airam.spiraali", "osram.classic.20w",
-               "airam.led.11w.4000k", "v.light.led.2w.6000k",
-               "osram.l36w.840", "osram.led.8w.2700k", "toshiba.led.9.5w.2700k",
-               "airam.led.oiva.9w.3000k", "osram.conc.spot.60w", 
-               "osram.led.10w.2700k.classicstar", "philips.tl5.35w.830he",
-               "pirkka.halogen.53w", "valoya.b50.ap67", 
-               "fluence.PhysioSpecGreenhouse",
-               "incandescent.60w", "osram.36w.25", "osram.hqit.400w",
-               "amaran.al.m9", "sunwayfoto.fl96.3000k", "sunwayfoto.fl96.5500k",
-               grep("^godox", names(lamps.mspct), value = TRUE) ))
+normalized.ls <- list()
+for (s in names(lamps.mspct)) {
+  normalized.ls[[s]] <- getNormalised(lamps.mspct[[s]])
+}
+peak.wl <- unlist(normalized.ls)
 
-## test vectors
-color_idxs <- 
-        unique(c(uv_lamps, amber_lamps, red_lamps, blue_lamps, white_lamps))
-color_idxs[!color_idxs %in% names(lamps.mspct)]
-stopifnot(all(color_idxs %in% names(lamps.mspct)))
+# as the spectra are normalised, a broad spectrum with have a large integral
+white_lamps <- lamps.mspct[q_irrad(lamps.mspct, allow.scaled = TRUE)[["Q_Total"]] > 2e-4 & 
+                             (peak.wl <= 700 | peak.wl > 890) & peak.wl >= 400]
+names(white_lamps)
+
+names_narrow <- setdiff(names(lamps.mspct), names(white_lamps))
+names_narrow
+
+peak.wl_narrow <- peak.wl[names_narrow]
+names_narrow == names(peak.wl_narrow)
+peak.wl_narrow <- unname(peak.wl_narrow)
+
+uv_lamps <- names_narrow[peak.wl_narrow <= wl_max(UV())]
+ir_lamps <- names_narrow[peak.wl_narrow > 700]
+purple_lamps <- names_narrow[peak.wl_narrow > wl_max(UV()) &
+                              peak.wl_narrow <= wl_max(Purple())]
+blue_lamps <- names_narrow[peak.wl_narrow > wl_min(Blue()) &
+                            peak.wl_narrow <= wl_max(Blue())]
+green_lamps <- names_narrow[peak.wl_narrow > wl_min(Green()) &
+                             peak.wl_narrow <= wl_max(Green())]
+yellow_lamps <- names_narrow[peak.wl_narrow > wl_min(Yellow()) &
+                              peak.wl_narrow <= wl_max(Yellow())]
+orange_lamps <- names_narrow[peak.wl_narrow > wl_min(Orange()) &
+                              peak.wl_narrow <= wl_max(Orange())]
+red_lamps <-  names_narrow[peak.wl_narrow > wl_min(Red()) &
+                            peak.wl_narrow <= wl_max(Red())]
+amber_lamps <- sort(c(yellow_lamps, orange_lamps))
+
+lamp_colors <- c("uv", "purle", "blue", "green", "yellow", "orange", "red", "ir")
 
 ## lists by type
 sodium_lamps <- "osram.super.vialox"
@@ -100,30 +106,20 @@ valoya_lamps <- sort(grep("valoya", names(lamps.mspct), value = TRUE))
 aputure_lamps <- sort(grep("amaran|aputure", names(lamps.mspct), value = TRUE))
 godox_lamps <- sort(grep("godox", names(lamps.mspct), value = TRUE))
 sunwayfoto_lamps <- sort(grep("sunwayfoto", names(lamps.mspct), value = TRUE))
+jaxman_lamps <- sort(grep("jaxman", names(lamps.mspct), value = TRUE))
+convoy_lamps <- sort(grep("convoy", names(lamps.mspct), value = TRUE))
 generic_lamps <- sort(grep("pirkka|v.light|LCDlighting", names(lamps.mspct), value = TRUE))
+
+lamp_brands <- 
+  sort(c("osram", "sylvania", "qpanel", "philips", "airam", "toshiba",
+         "valoya", "aputure", "godox", "sunwayfoto", "jaxman", "convoy",
+         "generic"))
 
 # lists by special uses
 plant_grow_lamps <- c(valoya_lamps, fluence_lamps)
 photography_lamps <- c(aputure_lamps, sunwayfoto_lamps, godox_lamps)
 
-## test vectors
-make_idxs <- unique(c(osram_lamps, sylvania_lamps, qpanel_lamps, philips_lamps, 
-                      airam_lamps, toshiba_lamps, valoya_lamps, fluence_lamps, 
-                      aputure_lamps, godox_lamps, sunwayfoto_lamps, generic_lamps))
-stopifnot(all(make_idxs %in% names(lamps.mspct)))
-
-## Test that all lamps are listed at least in one list of names
-message("Not in any index: ", 
-        setdiff(c(make_idxs, type_idxs, color_idxs),
-                names(lamps.mspct)))
-
-save(lamps.mspct, licor_lamps, macam_lamps, oo_maya_lamps, bentham_lamps,
-     uv_lamps, amber_lamps, red_lamps, blue_lamps, white_lamps,
-     sodium_lamps, incandescent_lamps, multimetal_lamps, mercury_lamps, led_lamps,
-     osram_lamps, sylvania_lamps, qpanel_lamps, philips_lamps,
-     airam_lamps, toshiba_lamps, generic_lamps, aputure_lamps, sunwayfoto_lamps,
-     godox_lamps, photography_lamps, xenon_lamps,
-     fluence_lamps, valoya_lamps, plant_grow_lamps, 
+save(list = c("lamps.mspct", "lamp_brands", "lamp_colors", ls(pattern = "*_lamps")),
      file = "data/lamps-mspct.rda")
 
 tools::resaveRdaFiles("data", compress="auto")